Extract CallsTable from CallsPage
